Point to my own csscomplete.vim repo
I have an open pull request that adds named colors and flex values into
the original repo. That PR has been sitting there for 6 months or so by
this point. Named colors are a convenience that I use all the time when
prototyping. Super important to me. So in the meantime, I will keep my
repo relatively up to date with the origin and layer my changes on top.