Allow selecting which packages to install
Depending on the installation location the installer may not want to
install all the listed packages. This makes the installer more
interactive by requiring a response for every package, but the tradeoff
is you can install minimal packages if you so choose.

You can still just hit (hold down) enter for every prompt and everything
would be installed as was previously the case.