Update init to no use svn
Previously I was using svn to sparsely checkout the install directory to
limit download times. Git didn't support doing this directly, but svn
came installed on new macs. This is not longer the case. Now git
supports it, and svn is not installed on new macs. Good thing they both
swapped, eh?