Update and add vim plugins (support neovim)
I've started using neovim, and it seems to be going well so far. I've attempted to have a
config that works for both vim and neovim depending on which is used.

This means I've swapped some plugins out for others that work better for both, or moved
them to optional starting and loading them conditionally.

Added:
* ToggleTerm
* CoC
* FzF
* vim-Plist

Update:
* vim-test

Removed:
* Ctrl-P

Moved:
* YouCompleteME