Update git config so I don't have uncommitted changes
Note: The generated files aren't committed and aren't going to exist.
And if you're already using this repo as your dotfiles (me on other
machines) they aren't going to be generated. But you should have
uncommitted changes and run into a conflict, so you can change it then.

User beware.