Git config updates
Mostly alias changes, but also added an optional termly config when in
the /tmp location.

I also opted out of the safe.directory check because I only ever use
this config in locations solely used by me, and any error it generates
is due to user differences that are still me.